fintrack/requirements.txt
Flask
render_template
request redirect
session
url_for
flash
mysql.connector
import generate_password_hash
check_password_hash